Added checks for edit and delete
Ensuring only admin and owner can perform the action

+++ forum/processes/post_process.py
@@ -4,18 +4,21 @@ from django.shortcuts import redirect
 from django.contrib import messages
 
 # MARK: Insert Post
-def process_create_post(request):
+def process_create_post(request, context={}):
     session_response = session_service.check_session(request)
 
     if session_response["status"] == "SUCCESS":
-        context = session_response["data"]
+        context["user_info"] = session_response["data"]
+    else:
+        messages.error(request, "Session Expired! Please login.")
+        return redirect("login_view")
 
     postTitle = request.POST.get("postTitle")
     postDescription = request.POST.get("postDescription")
     allowComments = request.POST.get("allowComments") == "on"
 
     response = post_service.insert_new_post(
-        postTitle, postDescription, allowComments, context["UserID"]
+        postTitle, postDescription, allowComments, context["user_info"]["UserID"]
     )
 
     if response["status"] == "SUCCESS":
@@ -24,33 +27,48 @@ def process_create_post(request):
     return redirect("index")
 
 # MARK: Delete Post by ID
-def process_delete_post(request, post_id):
+def process_delete_post(request, post_id, context={}):
     session_response = session_service.check_session(request)
 
     if session_response["status"] == "SUCCESS":
-        context = session_response["data"]
+        context["user_info"] = session_response["data"]
+    else:
+        messages.error(request, "Session Expired! Please login.")
+        return redirect("login_view")
+    
+    post_response = post_service.get_post_by_id(post_id)
+    if post_response["status"] == "SUCCESS":
+        context["post"] = post_response["data"]["post"]
 
-    response = post_service.delete_post_by_id(post_id)
+    if context["user_info"]["Role"] == "admin" or context["post"]["UserID_id"] == context["user_info"]["UserID"]:
+        response = post_service.delete_post_by_id(post_id)
 
-    if response["status"] == "SUCCESS":
-        return redirect('index')
+        if response["status"] == "SUCCESS":
+            return redirect('index')
 
     return redirect('index')
 
 # MARK: Update Post by ID
-def process_update_post(request, post_id):
+def process_update_post(request, post_id, context={}):
     session_response = session_service.check_session(request)
 
     if session_response["status"] == "SUCCESS":
-        context = session_response["data"]
+        context["user_info"] = session_response["data"]
+    else:
+        messages.error(request, "Session Expired! Please login.")
+        return redirect("login_view")
         
     postTitle = request.POST.get("postTitle")
     postDescription = request.POST.get("postDescription")
     allowComments = request.POST.get("allowComments") == "on"
 
-    response = post_service.update_post_by_id(postTitle, postDescription, allowComments, post_id)
+    post_response = post_service.get_post_by_id(post_id)
+    if post_response["status"] == "SUCCESS":
+        context["post"] = post_response["data"]["post"]
 
-    if response["status"] == "SUCCESS":
-        return redirect('post_view', post_id=post_id)
+    if context["user_info"]["Role"] == "admin" or context["post"]["UserID_id"] == context["user_info"]["UserID"]:        
+        response = post_service.update_post_by_id(postTitle, postDescription, allowComments, post_id)
+        if response["status"] == "SUCCESS":
+            return redirect('post_view', post_id=post_id)
 
     return redirect('post_view', post_id=post_id)
